fix: Update generateDB function to use Boolean values for signIn and signUp keys
This commit updates the generateDB function to use Boolean values (true) for signIn and signUp keys. The signIn and signUp fields used to be empty strings, which could be misleading when reading the code. Now they have a more explicit and cleaner meaning. Additionally, the 'collection' property is added to the user object to indicate which collection should be included.